+++ lib/pages/home_widgets/catalog_list.dart
@@ -1,0 +1,27 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_application_1/models/catalog.dart';
+import 'package:flutter_application_1/pages/home_details_page.dart';
+import 'package:flutter_application_1/pages/home_widgets/catalog_item.dart';
+
+class CatalogList extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return ListView.builder(
+        itemCount: CatalogModel.items.length,
+        itemBuilder: (context, index) {
+          final catalog = CatalogModel.items[index];
+          return InkWell(
+              onTap: () => {
+                    Navigator.push(
+                      context,
+                      MaterialPageRoute(
+                        builder: (context) => HomeDetails(
+                          catalog: catalog,
+                        ),
+                      ),
+                    )
+                  },
+              child: CatalogItem(catalog: catalog));
+        });
+  }
+}
