+++ lib/utils/MyRoutes.dart
@@ -1,4 +1,5 @@
 class MyRoutes {
   static String loginRoute = "/login";
   static String homeRoute = "/home";
+  static String detailsRoute = "/details";
 }
